smartpest_backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
Pillow==10.1.0
numpy==1.24.3 